Deleted old Settings components and moved new components out of v2 folder
