+++ src/components/settings-components/entries/CopyLogToClipboard.vue
@@ -1,14 +1,14 @@
 <script lang="ts" setup>
 import { computed, ref } from 'vue';
-import { getStore } from '../../../../providers/generic/store/StoreProvider';
-import { State } from '../../../../store';
+import { getStore } from '../../../providers/generic/store/StoreProvider';
+import { State } from '../../../store';
 import SettingsViewWrapper from '../SettingsViewWrapper.vue';
-import { useSettingSearch } from '../../../composables/SettingSearchComposable';
-import FsProvider from '../../../../providers/generic/file/FsProvider';
-import path from '../../../../providers/node/path/path';
-import InteractionProvider from '../../../../providers/ror2/system/InteractionProvider';
-import { PackageLoader } from '../../../../model/schema/ThunderstoreSchema';
-import R2Error from '../../../../model/errors/R2Error';
+import { useSettingSearch } from '../../composables/SettingSearchComposable';
+import FsProvider from '../../../providers/generic/file/FsProvider';
+import path from '../../../providers/node/path/path';
+import InteractionProvider from '../../../providers/ror2/system/InteractionProvider';
+import { PackageLoader } from '../../../model/schema/ThunderstoreSchema';
+import R2Error from '../../../model/errors/R2Error';
 
 const store = getStore<State>();
