+++ src/components/settings-components/entries/ModCache.vue
@@ -1,11 +1,11 @@
 <script lang="ts" setup>
 import { computed, ref } from 'vue';
-import { getStore } from '../../../../providers/generic/store/StoreProvider';
-import { State } from '../../../../store';
+import { getStore } from '../../../providers/generic/store/StoreProvider';
+import { State } from '../../../store';
 import SettingsViewWrapper from '../SettingsViewWrapper.vue';
-import { useSettingSearch } from '../../../composables/SettingSearchComposable';
-import CacheUtil from '../../../../r2mm/mods/CacheUtil';
-import R2Error from '../../../../model/errors/R2Error';
+import { useSettingSearch } from '../../composables/SettingSearchComposable';
+import CacheUtil from '../../../r2mm/mods/CacheUtil';
+import R2Error from '../../../model/errors/R2Error';
 
 const store = getStore<State>();
