+++ src/components/settings-components/entries/ToggleCdn.vue
@@ -1,8 +1,8 @@
 <script lang="ts" setup>
 import { ref } from 'vue';
 import SettingsViewWrapper from '../SettingsViewWrapper.vue';
-import { useSettingSearch } from '../../../composables/SettingSearchComposable';
-import CdnProvider from '../../../../providers/generic/connection/CdnProvider';
+import { useSettingSearch } from '../../composables/SettingSearchComposable';
+import CdnProvider from '../../../providers/generic/connection/CdnProvider';
 
 const props = defineProps<{
     searchTerm?: string;
