+++ src/providers/components/loaders/bind_impls/BindLoaderImpl.ts
@@ -5,7 +5,7 @@ import OnlineModListProvider from '../OnlineModListProvider';
 export default class BindLoaderImpl {
 
     static bind() {
-        SettingsViewProvider.provide(() => import("../../../../components/settings-components/v2/SettingsView.vue"));
+        SettingsViewProvider.provide(() => import("../../../../components/settings-components/SettingsView.vue"));
         LocalModListProvider.provide(() => import("../../../../components/views/LocalModList.vue"));
         OnlineModListProvider.provide(() => import("../../../../components/views/OnlineModListWithPanel.vue"));
     }
